+++ backend/tests/conftest.py
@@ -1,13 +1,11 @@
-import functools
 import logging
 import os
 from datetime import datetime, timedelta
-from typing import List, Union
 
 import pytest
 import pytz
 from fastapi.testclient import TestClient
-from sqlalchemy import inspect as sa_inspect, text
+from sqlalchemy import text
 from sqlmodel import Session, SQLModel, create_engine, select
 
 from backend.api import app
@@ -389,135 +387,58 @@ def setup_demo_data(db_session: Session) -> None:
     db_session.commit()
 
 
-def inspect_db_state(tables: Union[List[str], str] = None, all_tables: bool = False):
-    """
-    Decorator to inspect and print database state before and after a test.
-
-    Args:
-        tables: List of table names to inspect, or a single table name
-        all_tables: If True, inspect all tables regardless of tables parameter
-    """
-
-    def decorator(test_func):
-        @functools.wraps(test_func)
-        def wrapper(*args, **kwargs):
-            # Get db_session from kwargs or pytest fixture
-            session = None
-            for arg in list(kwargs.values()) + list(args):
-                if hasattr(arg, "execute") and hasattr(arg, "commit"):
-                    session = arg
-                    break
-
-            if not session:
-                print("WARNING: No database session found, cannot inspect db state")
-                return test_func(*args, **kwargs)
-
-            engine = session.get_bind()
-            inspector = sa_inspect(engine)
-
-            # Determine which tables to inspect
-            table_names = []
-            if all_tables:
-                table_names = inspector.get_table_names()
-            elif tables:
-                if isinstance(tables, str):
-                    table_names = [tables]
-                else:
-                    table_names = tables
-
-            print(f"\n=== DATABASE STATE BEFORE TEST: {test_func.__name__} ===")
-            for table_name in table_names:
-                print(f"\n--- TABLE: {table_name} ---")
-                try:
-                    result = session.execute(
-                        text(f"SELECT * FROM {table_name}")
-                    ).fetchall()
-                    if result:
-                        # Get column names
-                        columns = [
-                            col["name"] for col in inspector.get_columns(table_name)
-                        ]
-                        print(f"Columns: {columns}")
-                        print(f"Row count: {len(result)}")
-                        # Print first 5 rows for preview
-                        for i, row in enumerate(result[:5]):
-                            print(f"Row {i+1}: {row}")
-                        if len(result) > 5:
-                            print(f"... and {len(result) - 5} more rows")
-                    else:
-                        print("Table is empty")
-                except Exception as e:
-                    print(f"Error inspecting table {table_name}: {e}")
-
-            # Run the test
-            result = test_func(*args, **kwargs)
-
-            print(f"\n=== DATABASE STATE AFTER TEST: {test_func.__name__} ===")
-            for table_name in table_names:
-                print(f"\n--- TABLE: {table_name} ---")
-                try:
-                    result_after = session.execute(
-                        text(f"SELECT * FROM {table_name}")
-                    ).fetchall()
-                    if result_after:
-                        columns = [
-                            col["name"] for col in inspector.get_columns(table_name)
-                        ]
-                        print(f"Columns: {columns}")
-                        print(f"Row count: {len(result_after)}")
-                        for i, row in enumerate(result_after[:5]):
-                            print(f"Row {i+1}: {row}")
-                        if len(result_after) > 5:
-                            print(f"... and {len(result_after) - 5} more rows")
-                    else:
-                        print("Table is empty")
-                except Exception as e:
-                    print(f"Error inspecting table {table_name}: {e}")
-
-            return result
-
-        return wrapper
-
-    return decorator
+@pytest.fixture
+def admin_headers(auth_headers) -> dict:
+    """Alias for auth_headers — admin-role bearer headers."""
+    return auth_headers
 
 
-def print_db_state(session, tables=None, all_tables=False, label=""):
-    """
-    Function to inspect and print database state at any point in the code.
+@pytest.fixture
+def student_headers() -> dict:
+    """Generic student-role bearer headers (no team_id). Suitable for tests that
+    only check role-gating on admin-only endpoints."""
+    token = create_access_token(
+        data={"sub": "student", "role": "student"},
+        expires_delta=timedelta(minutes=30),
+    )
+    return {"Authorization": f"Bearer {token}"}
 
-    Args:
-        session: SQLAlchemy session
-        tables: List of table names to inspect, or a single table name
-        all_tables: If True, inspect all tables regardless of tables parameter
-        label: Optional label to include in the output
-    """
-    engine = session.get_bind()
-    inspector = sa_inspect(engine)
 
-    # Determine which tables to inspect
-    table_names = []
-    if all_tables:
-        table_names = inspector.get_table_names()
-    elif tables:
-        if isinstance(tables, str):
-            table_names = [tables]
-        else:
-            table_names = tables
+@pytest.fixture
+def team_headers(db_session) -> dict:
+    """TeamA seed-data bearer headers (role=student, includes team_id + institution_id)."""
+    team = db_session.exec(select(Team).where(Team.name == "TeamA")).first()
+    return {"Authorization": f"Bearer {make_student_token(team)}"}
 
-    print(f"\n=== DATABASE STATE {label} ===")
-    for table_name in table_names:
-        print(f"\n--- TABLE: {table_name} ---")
-        try:
-            result = session.execute(text(f"SELECT * FROM {table_name}")).fetchall()
-            if result:
-                columns = [col["name"] for col in inspector.get_columns(table_name)]
-                print(f"Columns: {columns}")
-                print(f"Row count: {len(result)}")
-                for i, row in enumerate(result[:5]):
-                    print(f"Row {i+1}: {row}")
-                if len(result) > 5:
-                    print(f"... and {len(result) - 5} more rows")
-            else:
-                print("Table is empty")
-        except Exception as e:
-            print(f"Error inspecting table {table_name}: {e}")
+
+@pytest.fixture
+def institution_token(db_session: Session) -> str:
+    """Create a fresh institution and return its institution-role JWT."""
+    institution = Institution(
+        name="test_institution",
+        contact_person="Test Person",
+        contact_email="test@example.com",
+        created_date=datetime.now(),
+        subscription_active=True,
+        subscription_expiry=datetime.now() + timedelta(days=30),
+        docker_access=True,
+        password_hash="test_hash",
+    )
+    db_session.add(institution)
+    db_session.commit()
+    db_session.refresh(institution)
+
+    return create_access_token(
+        data={
+            "sub": institution.name,
+            "role": "institution",
+            "institution_id": institution.id,
+        },
+        expires_delta=timedelta(minutes=30),
+    )
+
+
+@pytest.fixture
+def institution_headers(institution_token: str) -> dict:
+    """Bearer headers wrapping institution_token."""
+    return {"Authorization": f"Bearer {institution_token}"}
