+++ backend/tests/integration/test_game_workflows.py
@@ -15,9 +15,6 @@ from backend.database.db_models import (
     Team,
 )
 from backend.routes.auth.auth_core import create_access_token
-from backend.tests.conftest import inspect_db_state
-
-
 
 
 @pytest.fixture
@@ -51,7 +48,6 @@ class CustomPlayer(Player):
     return team
 
 
-@inspect_db_state(all_tables=True)
 def test_complete_game_lifecycle(
     client: TestClient,
     db_session: Session,
